Filling Feed View - get restaurant details TODO: send to table view of feed to occupy it
