+++ ForageApp/FeedView/FeedViewController.swift
@@ -14,18 +14,25 @@ class FeedViewController: UIViewController, UITableViewDelegate, UITableViewData
     @IBOutlet weak var NavBarCheckInBtn: UIBarButtonItem!
     @IBOutlet weak var FeedTableView: UITableView!
     
+    var viewModels = [RestaurantListViewModel]() {
+        didSet {
+            FeedTableView.reloadData()
+        }
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
         FeedTableView.dataSource = self
         FeedTableView.delegate = self
         
+        print(viewModels)
+        
 
     }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         // TODO: implement cell count
-        return 5
+        return viewModels.count
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
@@ -35,6 +42,9 @@ class FeedViewController: UIViewController, UITableViewDelegate, UITableViewData
         cell.restaurantImageView.layer.cornerRadius = 5;
         cell.restaurantImageView.layer.masksToBounds = true;
         // TODO: Setup FeedCell
+        let vm = viewModels[indexPath.row]
+        cell.configure(with: vm)
+        
         return cell
     }
