+++ ForageApp/NetworkService.swift
@@ -36,7 +36,7 @@ enum YelpService{
         var task: Task{
             switch self {
                 case let .search(lat, long):
-                    return .requestParameters(parameters: ["latitude": lat, "longitude": long, "limit": 1], encoding: URLEncoding.queryString)
+                    return .requestParameters(parameters: ["latitude": lat, "longitude": long, "limit": 10], encoding: URLEncoding.queryString)
             }
         }
